autoloom/pyproject.toml
[tool.poetry]
name = "lui"
version = "0.0.1"
description = "Loom User Interface"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "lui"}]

[tool.poetry.dependencies]
python = "^3.8"
textual = "^0.27.0"
openai = "^1.3.0"
rich = "^13.0.0"
aiohttp = "^3.8.0"
requests = "^2.32.3"
pyperclip = "^1.9.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
lui = "lui.main:main"
tuni = "tuni.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"